Created package for api docs
> Added .gitignore
> removed schema.py and placed the pydantic models in models.py
> database changes
