+++ .gitignore
@@ -1,3 +1,4 @@
-/__pycache__
-/env
+/env/
 .env
+**/__pycache__/
+*.pyc
