+++ api_docs/paths.py
@@ -1,0 +1,3 @@
+import os 
+
+room_view_doc = os.path.join('api_docs', 'room', 'view.html')
