+++ api_docs/room/view.html
@@ -1,0 +1,4 @@
+<body>
+    <h1>View Rooms</h1>
+    <h4>This route will return you 20 rooms after successfull request</h4>
+</body>
